complete support for multiple device types and devices
