+++ Directory.Build.targets
@@ -1,5 +1,89 @@
 <Project>
   
+  <!-- We copy over the LibTorch native binaries in order to run tests.  This can probably be automated better. -->
+  <!-- These lists are duplicated in redist/... with file splitting added for package prep --> 
+
+  <ItemGroup  Condition="'$(TestUsesLibTorch)' == 'true'">
+    <NativeAssemblyReference Include="LibTorchSharp" />
+  </ItemGroup>
+
+  <!-- Windows CPU libtorch binary list used for examples and testing -->
+  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64' and '$(OS)' == 'Windows_NT' and '$(TestUsesLibTorch)' == 'true' and '$(TestCuda)' != 'true'">
+    <NativeAssemblyReference Include="c10" />
+    <NativeAssemblyReference Include="asmjit" />
+    <NativeAssemblyReference Include="caffe2_detectron_ops" />
+    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
+    <NativeAssemblyReference Include="fbgemm" />
+    <NativeAssemblyReference Include="libiomp5md" />
+    <NativeAssemblyReference Include="libiompstubs5md" />
+    <NativeAssemblyReference Include="torch" />
+    <NativeAssemblyReference Include="torch_cpu" />
+    <NativeAssemblyReference Include="torch_global_deps" />
+  </ItemGroup>
+
+  <!-- Windows CUDA 10.2 libtorch binary list used for examples and testing -->
+  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64' and '$(OS)' == 'Windows_NT' and '$(TestUsesLibTorch)' == 'true' and '$(TestCuda)' == 'true'">
+    <NativeAssemblyReference Include="c10" Variant="cuda\" />
+    <NativeAssemblyReference Include="asmjit" Variant="cuda\" />
+    <NativeAssemblyReference Include="c10_cuda" Variant="cuda\" />
+    <NativeAssemblyReference Include="caffe2_detectron_ops_gpu" Variant="cuda\" />
+    <NativeAssemblyReference Include="caffe2_module_test_dynamic" Variant="cuda\" />
+    <NativeAssemblyReference Include="caffe2_nvrtc" Variant="cuda\" />
+    <NativeAssemblyReference Include="cublas64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="cublasLt64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="cudart64_102" Variant="cuda\" />
+    <NativeAssemblyReference Include="cudnn64_7" Variant="cuda\" />
+    <NativeAssemblyReference Include="cufft64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="cufftw64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="curand64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="cusparse64_10" Variant="cuda\" />
+    <NativeAssemblyReference Include="fbgemm" Variant="cuda\" />
+    <NativeAssemblyReference Include="libiomp5md" Variant="cuda\" />
+    <NativeAssemblyReference Include="libiompstubs5md" Variant="cuda\" />
+    <NativeAssemblyReference Include="nvrtc64_102_0" Variant="cuda\" />
+    <NativeAssemblyReference Include="nvrtc-builtins64_102" Variant="cuda\" />
+    <NativeAssemblyReference Include="nvToolsExt64_1" Variant="cuda\" />
+    <NativeAssemblyReference Include="torch" Variant="cuda\" />
+    <NativeAssemblyReference Include="torch_cpu" Variant="cuda\" />
+    <NativeAssemblyReference Include="torch_cuda" Variant="cuda\" />
+    <NativeAssemblyReference Include="torch_global_deps" Variant="cuda\" />
+  </ItemGroup>
+
+  <!-- Linux libtorch binary list used for examples and testing -->
+  <!-- NOTE: On Linux, examples and testing do not currently use Linux CUDA binaries -->
+  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64'and $([MSBuild]::IsOSPlatform('osx')) and '$(TestUsesLibTorch)' == 'true' ">
+    <NativeAssemblyReference Include="c10" />
+    <NativeAssemblyReference Include="caffe2_detectron_ops" />
+    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
+    <NativeAssemblyReference Include="caffe2_observers" />
+    <NativeAssemblyReference Include="fbjni" />
+    <NativeAssemblyReference Include="pytorch_jni" />
+    <NativeAssemblyReference Include="shm" />
+    <NativeAssemblyReference Include="torch" />
+    <NativeAssemblyReference Include="torch_cpu" />
+    <NativeAssemblyReference Include="torch_global_deps" />
+    <NativeAssemblyReference Include="torch_python" />
+    <NativeAssemblyReference Include="mklml" />
+    <NativeAssemblyReference Include="iomp5" />
+  </ItemGroup>
+
+  <!-- OSX CPU libtorch binary list used for examples and testing -->
+  <!-- NOTE: On OSX, examples and testing do not currently use Linux CUDA binaries -->
+  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64' and $([MSBuild]::IsOSPlatform('linux')) and '$(TestUsesLibTorch)' == 'true' and '$(TestCuda)' != 'true'">
+    <NativeAssemblyReference Include="c10" />
+    <NativeAssemblyReference Include="caffe2_detectron_ops" />
+    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
+    <NativeAssemblyReference Include="caffe2_observers" />
+    <NativeAssemblyReference Include="fbjni" />
+    <NativeAssemblyReference Include="pytorch_jni" />
+    <NativeAssemblyReference Include="shm" />
+    <NativeAssemblyReference Include="torch" />
+    <NativeAssemblyReference Include="torch_cpu" />
+    <NativeAssemblyReference Include="torch_global_deps" />
+    <NativeAssemblyReference Include="torch_python" />
+    <NativeAssemblyReference Include="gomp-7c85b1e2" ExtraExtension=".1" />
+  </ItemGroup>
+
   <Target Name="CheckForBuildTools">
     <Error Condition="!Exists('$(ToolsDir)')"
            Text="The tools directory [$(ToolsDir)] does not exist. Please run build in the root of the repo to ensure the tools are installed before attempting to build an individual project." />
@@ -17,7 +101,7 @@
 
     <ItemGroup>
       <NativeAssemblyReference>
-        <FullAssemblyPath>$(NativeOutputPath)$(NativeLibPrefix)%(NativeAssemblyReference.Identity)$(NativeLibExtension)%(NativeAssemblyReference.ExtraExtension)</FullAssemblyPath>
+        <FullAssemblyPath>$(NativeOutputPath)%(NativeAssemblyReference.Variant)$(NativeLibPrefix)%(NativeAssemblyReference.Identity)$(NativeLibExtension)%(NativeAssemblyReference.ExtraExtension)</FullAssemblyPath>
       </NativeAssemblyReference>
     </ItemGroup>
