+++ TorchSharp.sln
@@ -20,9 +20,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.ML.InternalCodeAn
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "common", "common", "{A84717CB-F11A-41C5-A74D-C0F1D47B7431}"
 	ProjectSection(SolutionItems) = preProject
+		pkg\common\DnnImageFeaturizer.props = pkg\common\DnnImageFeaturizer.props
 		pkg\common\RestitchPackage.props = pkg\common\RestitchPackage.props
 		pkg\common\RestitchPackage.targets = pkg\common\RestitchPackage.targets
-		pkg\common\DnnImageFeaturizer.props = pkg\common\DnnImageFeaturizer.props
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TorchSharp", "src\TorchSharp\TorchSharp.csproj", "{061CCBA1-A859-4392-8F45-249E5DAF1C88}"
@@ -48,11 +48,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Native-Debug", "Native-Debu
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Redist", "Redist", "{D8C60CD8-8429-45F2-A755-47B6CD10FDF8}"
 	ProjectSection(SolutionItems) = preProject
-		src\Redist\libtorch-cuda-10.2\libtorch-cuda-10.2.proj = src\Redist\libtorch-cuda-10.2\libtorch-cuda-10.2.proj
 		src\Redist\libtorch-cpu\libtorch-cpu.proj = src\Redist\libtorch-cpu\libtorch-cpu.proj
+		src\Redist\libtorch-cuda-10.2\libtorch-cuda-10.2.proj = src\Redist\libtorch-cuda-10.2\libtorch-cuda-10.2.proj
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Native-Release", "Native-Release", "{4DB9E84D-324C-408F-87A6-246E86205540}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TorchSharpTest.WithCudaBinaries", "test\TorchSharpTest.WithCudaBinaries\TorchSharpTest.WithCudaBinaries.csproj", "{DD652544-711E-4029-83FF-DA4A9600E6E7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -102,6 +104,14 @@ Global
 		{912E4543-FB5A-3172-BAFE-B3F15EE0A723}.Debug|x64.ActiveCfg = Debug|x64
 		{912E4543-FB5A-3172-BAFE-B3F15EE0A723}.Release|Any CPU.ActiveCfg = Release|x64
 		{912E4543-FB5A-3172-BAFE-B3F15EE0A723}.Release|x64.ActiveCfg = Release|x64
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Debug|x64.Build.0 = Debug|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Release|x64.ActiveCfg = Release|Any CPU
+		{DD652544-711E-4029-83FF-DA4A9600E6E7}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -118,6 +128,7 @@ Global
 		{CF2C1A9E-3A8A-4329-8A6E-7880C15AAC3D} = {09EADF06-BE25-4228-AB53-95AE3E15B530}
 		{D8C60CD8-8429-45F2-A755-47B6CD10FDF8} = {09EADF06-BE25-4228-AB53-95AE3E15B530}
 		{4DB9E84D-324C-408F-87A6-246E86205540} = {09EADF06-BE25-4228-AB53-95AE3E15B530}
+		{DD652544-711E-4029-83FF-DA4A9600E6E7} = {AED9C836-31E3-4F3F-8ABC-929555D3F3C4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {41165AF1-35BB-4832-A189-73060F82B01D}
