+++ src/Examples/Examples.csproj
@@ -3,63 +3,16 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TestUsesLibTorch>true</TestUsesLibTorch>
     <UseMLCodeAnalyzer>false</UseMLCodeAnalyzer>
     <UseStyleCopAnalyzer>false</UseStyleCopAnalyzer>
     <StartupObject>TorchSharp.Examples.AlexNet</StartupObject>
-      <IsPackable>false</IsPackable>
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\TorchSharp\TorchSharp.csproj" />
   </ItemGroup>
   
-  <ItemGroup>
-    <NativeAssemblyReference Include="LibTorchSharp" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64' and '$(OS)' == 'Windows_NT'">
-    <NativeAssemblyReference Include="c10" />
-    <NativeAssemblyReference Include="asmjit" />
-    <NativeAssemblyReference Include="caffe2_detectron_ops" />
-    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
-    <NativeAssemblyReference Include="fbgemm" />
-    <NativeAssemblyReference Include="libiomp5md" />
-    <NativeAssemblyReference Include="libiompstubs5md" />
-    <NativeAssemblyReference Include="torch" />
-    <NativeAssemblyReference Include="torch_cpu" />
-    <NativeAssemblyReference Include="torch_global_deps" />
-  </ItemGroup>
-
-    <!-- Torch is 64-bit only? REVIEW check if true do we need all these? -->
-  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64'and $([MSBuild]::IsOSPlatform('osx'))">
-    <NativeAssemblyReference Include="c10" />
-    <NativeAssemblyReference Include="caffe2_detectron_ops" />
-    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
-    <NativeAssemblyReference Include="caffe2_observers" />
-    <NativeAssemblyReference Include="fbjni" />
-    <NativeAssemblyReference Include="pytorch_jni" />
-    <NativeAssemblyReference Include="shm" />
-    <NativeAssemblyReference Include="torch" />
-    <NativeAssemblyReference Include="torch_cpu" />
-    <NativeAssemblyReference Include="torch_global_deps" />
-    <NativeAssemblyReference Include="torch_python" />
-    <NativeAssemblyReference Include="mklml" />
-    <NativeAssemblyReference Include="iomp5" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(NativeTargetArchitecture)' == 'x64' and $([MSBuild]::IsOSPlatform('linux'))">
-    <NativeAssemblyReference Include="c10" />
-    <NativeAssemblyReference Include="caffe2_detectron_ops" />
-    <NativeAssemblyReference Include="caffe2_module_test_dynamic" />
-    <NativeAssemblyReference Include="caffe2_observers" />
-    <NativeAssemblyReference Include="fbjni" />
-    <NativeAssemblyReference Include="pytorch_jni" />
-    <NativeAssemblyReference Include="shm" />
-    <NativeAssemblyReference Include="torch" />
-    <NativeAssemblyReference Include="torch_cpu" />
-    <NativeAssemblyReference Include="torch_global_deps" />
-    <NativeAssemblyReference Include="torch_python" />
-    <NativeAssemblyReference Include="gomp-7c85b1e2" ExtraExtension=".1" />
-  </ItemGroup>
 
 </Project>
