+++ src/TorchSharp/Tensor/TorchTensorTyped.tt
@@ -37,7 +37,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor Arange(TorchScalar start, TorchScalar stop, TorchScalar step, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             return new TorchTensor (THSTensor_arange (start.Handle, stop.Handle, step.Handle, (sbyte)ScalarType.<#=type.Name#>, (int) deviceType, deviceIndex, requiresGrad));
         }
@@ -50,7 +50,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor RandomPermutation(long n, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             return new TorchTensor (THSTensor_randperm (n, (sbyte)ScalarType.<#=type.Name#>, (int) deviceType, deviceIndex, requiresGrad));
         }
@@ -63,7 +63,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor Zeros(long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -82,7 +82,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor Ones(long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -101,7 +101,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor Empty(long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -120,7 +120,7 @@ foreach (var type in TorchTypeDef.Types) {
         /// </summary>
         static public TorchTensor RandomIntegers(long max, long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -141,7 +141,7 @@ if (type.IsFloatingPoint) {
         /// </summary>
         static public TorchTensor Random(long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -160,7 +160,7 @@ if (type.IsFloatingPoint) {
         /// </summary>
         static public TorchTensor RandomN(long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
@@ -249,7 +249,7 @@ if (type.IsLong) {
 
         public static TorchTensor Sparse(TorchTensor indices, TorchTensor values, long[] size, DeviceType deviceType = DeviceType.CPU, int deviceIndex = 0, bool requiresGrad = false)
         {
-            TorchTensor.CheckForCUDA (deviceType, deviceIndex);
+            Torch.InitializeDevice (deviceType, deviceIndex);
 
             unsafe
             {
