+++ test/TorchSharpTest.WithCudaBinaries/TorchSharpTest.WithCudaBinaries.csproj
@@ -1,8 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.0</TargetFramework>
     <TestUsesLibTorch>true</TestUsesLibTorch>
+    <TestCuda>true</TestCuda>
     <IsPackable>false</IsPackable>
     <PlatformTarget>x64</PlatformTarget>
     <UseStyleCopAnalyzer>false</UseStyleCopAnalyzer>
@@ -10,6 +11,13 @@
     <VSTestResultsDirectory>$(OutputPath)</VSTestResultsDirectory>
     <AssemblyOriginatorKeyFile>$(ToolsDir)Test.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="..\TorchSharpTest\GlobalSuppressions.cs" Link="GlobalSuppressions.cs" />
+    <Compile Include="..\TorchSharpTest\NN.cs" Link="NN.cs" />
+    <Compile Include="..\TorchSharpTest\TorchSharp.cs" Link="TorchSharp.cs" />
+    <Compile Include="..\TorchSharpTest\TorchTensor.cs" Link="TorchTensor.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\TorchSharp\TorchSharp.csproj" />
@@ -26,19 +34,6 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
   </ItemGroup>
 
-
-  <PropertyGroup Condition="'$(Coverage)' == 'true'">
-      <CollectCoverage>true</CollectCoverage>
-      <SingleHit>true</SingleHit>
-      <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies> <!-- https://github.com/tonerdo/coverlet/issues/72 -->
-      <CoverletOutputFormat>opencover</CoverletOutputFormat>
-      <CoverletOutput>$(BaseOutputPath)$(PlatformConfig)\coverage\$(MSBuildProjectName).coverage</CoverletOutput>
-      <Include>[Microsoft.ML.*]*</Include>
-      <!-- Excluding for perf reasons. These classes have tests that can be run conditionally
-           but they need to be migrated. Excluding these classes should have very minimal effect on code coverage.
-      -->
-      <ExcludeByAttribute>Obsolete,ExcludeFromCodeCoverage</ExcludeByAttribute>
-  </PropertyGroup>
 
 </Project>
